refactor classifiers models, add testcases for new models
