+++ kashgari/tasks/base/base_model.py
@@ -23,6 +23,7 @@ from kashgari.utils import helper
 from kashgari.embeddings import CustomEmbedding, BaseEmbedding
 from kashgari.utils.crf import CRF, crf_loss, crf_accuracy
 from keras_bert.bert import get_custom_objects as get_bert_custom_objects
+from kashgari.layers import AttentionWeightedAverage, KMaxPooling
 
 
 class BaseModel(object):
@@ -94,7 +95,8 @@ class BaseModel(object):
         if embedding and embedding['embedding_type'] == 'bert':
             custom_objects['NonMaskingLayer'] = helper.NonMaskingLayer
             custom_objects.update(get_bert_custom_objects())
-
+        custom_objects['AttentionWeightedAverage'] = AttentionWeightedAverage
+        custom_objects['KMaxPooling'] = KMaxPooling
         return custom_objects
 
     @classmethod
